InvoiceParserHTTP/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
lxml
pandas
azurebatchload
pyodbc
bs4
sqlalchemy
azure-keyvault-secrets
azure-identity